requirements.txt
hikari == 2.0.0.dev124 
hikari-lightbulb == 2.3.5
ossapi == 3.4.4